Fix and tests to avoid registered users being invited again
