+++ decidim-participatory_processes/spec/commands/create_participatory_process_admin_spec.rb
@@ -97,6 +97,13 @@ module Decidim::ParticipatoryProcesses
         expect(user.follows?(my_process)).to be true
       end
 
+      it "doesn't invite the user again" do
+        subject.call
+        user.reload
+
+        expect(user.invited_to_sign_up?).not_to be true
+      end
+
       context "when there is no user with the given email" do
         let(:email) { "does_not_exist@example.com" }
